moved math.h out of its own directory (there was only one file in the directory.
